fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,13 +9,13 @@
 
 int main(void)
 {
-	int n1 = 0, n2 = 1, n3, c = 0;
+	unsigned long n1 = 0, n2 = 1, n3, c = 0;
 
 	while (c <= 50)
 	{
 		n3 = n2 + n1;
 		n1 = n2;
-		printf("%d, ", n3);
+		printf("%lu, ", n3);
 		n2 = n3;
 		c++;
 	}
